% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_limit}
\alias{add_limit}
\title{add limit n in SQL query}
\usage{
add_limit(query, n = 100)
}
\arguments{
\item{query}{SQL statement}

\item{n}{number of observations}
}
\value{
query with added limit
}
\description{
add limit n in SQL query
}
\examples{
library(magrittr)
"select * from schema.table;" \%>\% add_limit(100)
}
