% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_paren}
\alias{add_paren}
\title{Concatenate all elements and surround the result by ()}
\usage{
add_paren(x)
}
\arguments{
\item{x}{vector}
}
\value{
a character string
}
\description{
Concatenate the elements in x into a string and surround the result by parentheses.
This does not add single quote to each element
}
\examples{
paren:::add_paren(1:5)
}
