% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{comma_concat}
\alias{comma_concat}
\title{Concatenate by comma}
\usage{
comma_concat(vars)
}
\arguments{
\item{vars}{character vector}
}
\value{
string with comma separated
}
\description{
Concatenate by comma
}
