% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_from}
\alias{count_from}
\title{Count the cases for the combination of variables from the table}
\usage{
count_from(vars, table_name)
}
\arguments{
\item{vars}{character vector of variable names}

\item{table_name}{string of table name}
}
\value{
sql query string
}
\description{
Count the cases for the combination of variables from the table
}
\examples{
library(tidyverse)
c("x", "y") \%>\%
count_from("table_name")
}
