% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_rows}
\alias{count_rows}
\title{Count number of rows in a table}
\usage{
count_rows(table_name)
}
\arguments{
\item{table_name}{table name}
}
\value{
SQL query of counting the number of rows
}
\description{
Count number of rows in a table
}
\examples{
count_rows("schema.table")
}
