% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paren}
\alias{paren}
\title{Surround the elements in parenthesis with optional single quotes}
\usage{
paren(x)
}
\arguments{
\item{x}{vector}
}
\value{
a character string
}
\description{
Mainly to use with SQL IN operator.
Surround the elements in parenthesis.
Each element will be single quoted depending on the data type.
}
\examples{
paren(1:5)
paren(letters[1:5])
# Use as follows to build SQL queries
paste("select * from schema.table where id in", paren(1:5), ";")
paste("select * from schema.table where id in", paren(letters[1:6]), ";")
}
