% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_semicolon}
\alias{remove_semicolon}
\title{Remove semicolon}
\usage{
remove_semicolon(x)
}
\arguments{
\item{x}{string or query}
}
\value{
string without semicolon
}
\description{
Remove semicolon
}
