% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{single_quote_each}
\alias{single_quote_each}
\title{Put each element in single quotes}
\usage{
single_quote_each(x)
}
\arguments{
\item{x}{vector}
}
\value{
a character string
}
\description{
Put each element in single quotes
}
\examples{
paren:::single_quote_each(letters[1:5])
}
